+++ CHANGELOG.md
@@ -1,7 +1,30 @@
-v1.00
+# CLICKERY HEXAGON CHANGELOG
+
+1.0.0
 - The game released, with its ascension and everything and its bugs, so crazy
 
-v1.01
+1.0.1
 - Game didn't ping newgrounds so leaderboards wouldn't work and medals i think
 - Fixed songs that were sped up due to audacity tempo stuff
 - Added more blab phrases
+
+1.1.0
+FEATURES:
+- Added stat for game complete time
+
+FIXES/CHANGES:
+- Removed position setters
+- Modified formula for getting prices after ascending
+- Hopefully made saves compatible with previous ones in case they don't have the properties
+- Hopefully fix the issue with medals, newgrounds wasn't being enabled on newgrounds (ironically)
+- Now at log in, for every medal you have in game but not on newgrounds, it will be unlocked every 5 seconds each
+
+- Changed critical chance
+- Goober devky has more chance of appearing
+- Gnome is SLIGHTLY rarer
+- Fixed formatting:
+	- On the ascendWin mana counter
+	- Mana per second on ascendWin
+	- Ascension substate mana counter
+	- Store element power text
+- Fixed store element for powerups not showing the correct price after ascending